requirements.txt
pandas==2.3.1
numpy==2.2.6
scikit-learn==1.6.1
matplotlib==3.10.7
seaborn==0.13.2
xgboost==3.1.1
joblib==1.4.2
streamlit==1.41.1
requests==2.32.3
python-dotenv==1.0.0